requirements.txt
asgiref==3.8.1
click==8.1.7
cloudinary==1.36.0
dj-database-url==0.5.0
dj-rest-auth==6.0.0
dj3-cloudinary-storage==0.0.6
Django==5.0.6
django-allauth==64.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.94.1
gunicorn==22.0.0
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
svgwrite==1.4.3
Tree==0.2.4
urllib3==1.26.18
whitenoise==6.6.0
